Rename _dash_files files
