+++ vizro-core/src/vizro/tables/__init__.py
@@ -1,5 +1,5 @@
-from vizro.tables.dash_aggrid import dash_ag_grid
-from vizro.tables.dash_table import dash_data_table
+from vizro.tables._dash_ag_grid import dash_ag_grid
+from vizro.tables._dash_table import dash_data_table
 
 # Please keep alphabetically ordered
 __all__ = ["dash_ag_grid", "dash_data_table"]
